+++ tests/2_models/model_smoke/test_wake_superpositions.py
@@ -10,14 +10,10 @@ MODEL_PATHS = [
     "foxes.models.wake_superpositions.ti_quadratic:TIQuadratic",
     "foxes.models.wake_superpositions.wind_vector:WindVectorLinear",
     "foxes.models.wake_superpositions.ws_linear:WSLinear",
-    "foxes.models.wake_superpositions.ws_linear:WSLinearLocal",
     "foxes.models.wake_superpositions.ws_max:WSMax",
-    "foxes.models.wake_superpositions.ws_max:WSMaxLocal",
     "foxes.models.wake_superpositions.ws_pow:WSPow",
-    "foxes.models.wake_superpositions.ws_pow:WSPowLocal",
     "foxes.models.wake_superpositions.ws_product:WSProduct",
     "foxes.models.wake_superpositions.ws_quadratic:WSQuadratic",
-    "foxes.models.wake_superpositions.ws_quadratic:WSQuadraticLocal",
 ]